+++ osu.Game/Screens/Multi/Play/TimeshiftPlayer.cs
@@ -14,9 +14,7 @@ using osu.Game.Online.API.Requests;
 using osu.Game.Online.Multiplayer;
 using osu.Game.Rulesets;
 using osu.Game.Scoring;
-using osu.Game.Screens.Multi.Ranking;
 using osu.Game.Screens.Play;
-using osu.Game.Screens.Ranking;
 
 namespace osu.Game.Screens.Multi.Play
 {
@@ -115,7 +113,5 @@ namespace osu.Game.Screens.Multi.Play
 
             Exited = null;
         }
-
-        protected override Results CreateResults(ScoreInfo score) => new MatchResults(score);
     }
 }
