--- osu.Game/Screens/Play/SoloResults.cs
@@ -1,24 +1,0 @@
-// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-using System.Collections.Generic;
-using osu.Game.Scoring;
-using osu.Game.Screens.Ranking;
-using osu.Game.Screens.Ranking.Types;
-
-namespace osu.Game.Screens.Play
-{
-    public class SoloResults : Results
-    {
-        public SoloResults(ScoreInfo score)
-            : base(score)
-        {
-        }
-
-        protected override IEnumerable<IResultPageInfo> CreateResultPages() => new IResultPageInfo[]
-        {
-            new ScoreOverviewPageInfo(Score, Beatmap.Value),
-            new LocalLeaderboardPageInfo(Score, Beatmap.Value)
-        };
-    }
-}
